refactor(parameters): * removed all local parameter utilities that are also present in @jscad/core * renamed getParameterValues for clarity and to match @jscad/core * updated all imports / uses accordingly to the points above
